fix macOS build: platform guards, includes, and codegen workarounds
- make.vsh: skip libtcc1 runtime on macOS (TCC can't target ARM64),
  use -cc cc -gc none -Wno-int-conversion for macOS builds
- bt-exe.v: wrap sigaction/signal handling in $if linux
- tcc.h.v: add string.h/stdlib.h/stdio.h includes for clang on macOS
- tccgen.v: workaround V address-taken variable codegen bug
- tccrun.v: add @[typedef] to duplicate va_list declaration

+++ make.vsh
@@ -2,16 +2,24 @@ import os
 
 fn build_steps(vtccdir string) ! {
 	mut deps := map[string]string{}
-	deps['building bt-log.o (tcc_backtrace symbol)'] = 'v -cg -keepc -d no_main -skip-unused -no-builtin -o bt-log.o lib/bt-log.v'
-	deps['building bcheck.o: '] = 'v -d no_main -skip-unused -no-builtin -o bcheck.o lib/bcheck.v'
-	deps['building bt-exe.o: '] = 'v -keepc -d no_main -skip-unused -no-builtin -o bt-exe.o lib/bt-exe.v'
-	deps['building dso.o: '] = 'v -d no_main -skip-unused -no-builtin -o dso.o lib/dsohandle.v'
-	deps['buildind libtcc1.o: '] = 'v -d no_main -no-builtin -o libtcc1.o lib/libtcc1/'
-	deps['building vtcc.o: '] = 'v -d vtcc_dir=${vtccdir} -d no_main -skip-unused -o vtcc.o src/'
-	deps['building libtcc1.a: '] = 'rm -f libtcc1.a && ar rcs libtcc1.a libtcc1.o bt-exe.o bt-log.o bcheck.o dso.o'
-	deps['buildind vtcc: '] = 'v -cg -skip-unused -d vtcc_dir=${vtccdir} -o vtcc src/'
+	cc_flag := if os.user_os() == 'macos' { ' -cc cc' } else { '' }
+	gc_flag := if os.user_os() == 'macos' { ' -gc none' } else { '' }
+	wflag := if os.user_os() == 'macos' { " -cflags '-Wno-int-conversion'" } else { '' }
+	if os.user_os() != 'macos' {
+		deps['building bt-log.o (tcc_backtrace symbol)'] = 'v -cg -keepc -d no_main -skip-unused -no-builtin -o bt-log.o lib/bt-log.v'
+		deps['building bcheck.o: '] = 'v -d no_main -skip-unused -no-builtin -o bcheck.o lib/bcheck.v'
+		deps['building bt-exe.o: '] = 'v -keepc -d no_main -skip-unused -no-builtin -o bt-exe.o lib/bt-exe.v'
+		deps['building dso.o: '] = 'v -d no_main -skip-unused -no-builtin -o dso.o lib/dsohandle.v'
+		deps['buildind libtcc1.o: '] = 'v -d no_main -no-builtin -o libtcc1.o lib/libtcc1/'
+		deps['building vtcc.o: '] = 'v -d vtcc_dir=${vtccdir} -d no_main -skip-unused -o vtcc.o src/'
+		deps['building libtcc1.a: '] = 'rm -f libtcc1.a && ar rcs libtcc1.a libtcc1.o bt-exe.o bt-log.o bcheck.o dso.o'
+	} else {
+		deps['building vtcc.o: '] = 'v -d vtcc_dir=${vtccdir} -d no_main -skip-unused${cc_flag}${gc_flag}${wflag} -o vtcc.o src/'
+	}
+	cg_flag := if os.user_os() == 'macos' { '' } else { ' -cg' }
+	deps['buildind vtcc: '] = 'v${cg_flag} -skip-unused -d vtcc_dir=${vtccdir}${cc_flag}${gc_flag}${wflag} -o vtcc src/'
 
-	os.execute('rm -f *.{a,o}')
+	os.execute('bash -c "rm -f *.a *.o"')
 
 	for k, v in deps {
 		eprint('${k}: ')
